refactor(validator/naming): extract naming CLI helpers into naming/src/cli
